gradio_demo/requirements.txt
ultralytics
huggingface_hub
gradio
spaces
opencv-python-headless
Pillow
numpy
torch
